# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI 模型提供商管理服务 API 文档\n提供模型提供商、模型信息和参数规则的查询接口",
        "title": "Genkit AI Service API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/audit/auth": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查询认证审计日志，支持多条件过滤和分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit"
                ],
                "summary": "查询审计日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "事件类型（login, logout, refresh, revoke, failed_login）",
                        "name": "event",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始时间（RFC3339格式）",
                        "name": "startTime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间（RFC3339格式）",
                        "name": "endTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/AuthAuditListResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/chat/abort": {
            "post": {
                "description": "中止指定消息的对话处理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "中止对话",
                "parameters": [
                    {
                        "description": "中止请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AbortRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功中止对话",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "消息不存在",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "AbortRequest": {
            "type": "object",
            "required": [
                "messageId"
            ],
            "properties": {
                "messageId": {
                    "description": "消息ID（必填）",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "AuthAuditItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "事件发生时间",
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                },
                "event": {
                    "description": "事件类型",
                    "type": "string",
                    "example": "login"
                },
                "id": {
                    "description": "审计日志ID",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "ip": {
                    "description": "客户端IP地址",
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "meta": {
                    "description": "事件元数据"
                },
                "tenantId": {
                    "description": "租户ID",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "userAgent": {
                    "description": "用户代理字符串",
                    "type": "string",
                    "example": "Mozilla/5.0"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "AuthAuditListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "分页数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/PaginationData-array_model_AuthAuditItem"
                        }
                    ]
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "请求参数错误"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
...
        "PaginationData-array_model_AuthAuditItem": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/AuthAuditItem"
                    }
                },
                "pageNo": {
                    "description": "当前页码",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "每页大小",
                    "type": "integer",
                    "example": 10
                },
                "totalCount": {
                    "description": "总记录数",
                    "type": "integer",
                    "example": 100
                },
                "totalPage": {
                    "description": "总页数",
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "操作成功"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer {token}\" 格式的 JWT 令牌进行身份认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}